+++ apps/storybook-react-native/babel.config.js
@@ -3,7 +3,7 @@ module.exports = function (api) {
   return {
     presets: ['babel-preset-expo'],
     plugins: [
-      'react-native-reanimated/plugin', // Must be the last plugin
+      'react-native-worklets/plugin', // Must be the last plugin
     ],
   };
 };