+++ packages/design-system-react-native/src/components/BottomSheetDialog/BottomSheetDialog.tsx
@@ -22,7 +22,6 @@ import {
 } from 'react-native';
 import { Gesture, GestureDetector } from 'react-native-gesture-handler';
 import Animated, {
-  runOnJS,
   useAnimatedStyle,
   useSharedValue,
   withTiming,
@@ -31,6 +30,7 @@ import {
   useSafeAreaFrame,
   useSafeAreaInsets,
 } from 'react-native-safe-area-context';
+import { scheduleOnRN } from 'react-native-worklets';
 
 // Internal dependencies.
 import {
@@ -95,9 +95,9 @@ export const BottomSheetDialog = forwardRef<
           bottomOfDialogYValue.value,
           { duration: DEFAULT_BOTTOMSHEETDIALOG_DISPLAY_DURATION },
           () => {
-            runOnJS(onCloseCB)();
+            scheduleOnRN(onCloseCB);
             if (callback) {
-              runOnJS(callback)();
+              scheduleOnRN(callback);
             }
           },
         );
@@ -177,7 +177,7 @@ export const BottomSheetDialog = forwardRef<
           const isDismissed = finalYOffset === bottomOfDialogYValue.value;
 
           if (isDismissed) {
-            runOnJS(onCloseDialog)();
+            scheduleOnRN(onCloseDialog);
           } else {
             // Only animate dialog to a certain Y position instead
             currentYOffset.value = withTiming(finalYOffset, {
@@ -207,9 +207,9 @@ export const BottomSheetDialog = forwardRef<
           duration: DEFAULT_BOTTOMSHEETDIALOG_DISPLAY_DURATION,
         },
         () => {
-          runOnJS(onOpenCB)();
+          scheduleOnRN(onOpenCB);
           if (callback) {
-            runOnJS(callback)();
+            scheduleOnRN(callback);
           }
         },
       );