+++ packages/design-system-react-native/src/components/HelpText/index.ts
@@ -1,0 +1,3 @@
+export { HelpTextSeverity } from '@metamask/design-system-shared';
+export { HelpText } from './HelpText';
+export type { HelpTextProps } from './HelpText.types';